pipeline/requirements.txt
requests
pylint
pytest
pyodbc
pymssql
python-dotenv
boto3